+++ gtfstk/constants.py
@@ -65,3 +65,10 @@ INT_COLS = [
 
 #:
 DISTANCE_UNITS = ['ft', 'mi', 'm', 'km']
+
+#:
+CRS_WGS84 = {'no_defs': True, 'ellps': 'WGS84', 'datum': 
+  'WGS84', 'proj': 'longlat'}
+
+#:
+BIG = 1000000